cobaya2: isolating yaml dependency
